gherkin: (C) Handle multi language filenames on Windows.
The Gherkin library assumes that the feature filenames are encoded in
UTF-8. On Windows that most likely is not true. Even if Windows are set
to use unicode, that usually is UTF-16 and not UTF-8. Therefore, on
Windows, use a Windows API to convert from the encoding of the CLI
arguments (feature filename) and the internal wide character
representation.